+++ tests/spec/failed-tests.txt
@@ -1,5 +1,3 @@
-FAIL  tests/spec/spec-update.dont-add-comma.spec.ts
-
 FAIL  tests/update-custom/update-custom-mark-case-insensitive-and-methods-inside-describe.spec.ts
 
 FAIL  tests/update-custom/update-custom-green-path.spec.ts
