+++ tests/spec/test-data/update-no-comma/c.ts
@@ -1,0 +1,5 @@
+import { LogService, BDep } from '@angular/core';
+
+export class C {
+    constructor(private bDep: bDep, private logger: LogService) {}
+}
